+++ pyannote/audio/utils/frame.py
@@ -1,0 +1,80 @@
+# MIT License
+#
+# Copyright (c) 2023 CNRS
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in all
+# copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+# SOFTWARE.
+
+
+import math
+
+
+def conv1d_num_frames(num_samples, kernel_size=5, stride=1, padding=0, dilation=1):
+    """Compute expected number of frames after 1D convolution
+
+    Parameters
+    ----------
+    num_samples : int
+        Number of samples in the input signal
+    kernel_size : int
+        Kernel size
+    stride : int
+        Stride
+    padding : int
+        Padding
+    dilation : int
+        Dilation
+
+    Returns
+    -------
+    num_frames : int
+        Number of frames in the output signal
+
+    Source
+    ------
+    https://pytorch.org/docs/stable/generated/torch.nn.Conv1d.html#torch.nn.Conv1d
+    """
+    return math.floor(
+        1 + (num_samples + 2 * padding - dilation * (kernel_size - 1) - 1) / stride
+    )
+
+
+def conv1d_receptive_field_size(
+    num_frames=1, kernel_size=5, stride=1, padding=0, dilation=1
+):
+    """Compute receptive field size for `num_frames` frames after 1D convolution
+
+    Parameters
+    ----------
+    num_frames : int, optional
+        Number of frames in the output signal
+    kernel_size : int
+        Kernel size
+    stride : int
+        Stride
+    padding : int
+        Padding
+    dilation : int
+        Dilation
+
+    Returns
+    -------
+    receptive_field : int
+        Receptive field size
+    """
+    return (num_frames - 1) * stride - 2 * padding + dilation * (kernel_size - 1) + 1