+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## develop branch
+
+### New features
+
+- feat(model): add `num_frames` and `receptive_field` to segmentation models
+
 ## Version 3.1.1 (2023-12-01)
 
 ### TL;DR